remove center store module
