--- app/web/store/lesson/center/getters.js
@@ -1,6 +1,0 @@
-
-const getters = {
-  packagesList: (state) => state.packagesList
-}
-
-export default getters
