--- app/web/store/lesson/center/index.js
@@ -1,15 +1,0 @@
-import getters from './getters'
-import actions from './actions'
-import mutations from './mutations'
-
-const state = () => ({
-  packagesList: {}
-})
-
-export default {
-  namespaced: true,
-  state,
-  getters,
-  actions,
-  mutations
-}
